refactor: resolve conflict
