+++ apps/designer/globalStore/Navigation-Store/navigationStore.ts
@@ -1,0 +1,33 @@
+import create from "zustand";
+type navigationStore = {
+  isNavbarOpen: boolean;
+  handleClickOnWelcome: () => void;
+  isWelcomePanelActive: boolean;
+  isToggled: boolean;
+  handleClickOnNewButton: () => void;
+  handleClickForToggle: () => void;
+};
+const navigationUseStore = create<navigationStore>((set) => ({
+  isWelcomePanelActive: false, //for welcome component
+  isNavbarOpen: false, //for navbar by default is closed
+  handleClickOnWelcome: () => {
+    set(() => ({
+      isNavbarOpen: true,
+      isWelcomePanelActive: true,
+    }));
+  },
+  isToggled: false,
+  handleClickOnNewButton: () => {
+    set(() => ({
+      isWelcomePanelActive: false,
+      isNavbarOpen: false,
+      isToggled: false,
+    }));
+  },
+  handleClickForToggle: () => {
+    set((state) => ({
+      isToggled: !state.isToggled,
+    }));
+  },
+}));
+export default navigationUseStore;
