+++ apps/designer/globalStore/Navigation-Store/openWelcomeStore.ts
@@ -1,0 +1,13 @@
+import create from "zustand";
+type openWelcomeStore = {
+  isWelcomeComponentOpen: boolean;
+  handleClickOnNewDesign: () => void;
+};
+export const openWelcomeStore = create<openWelcomeStore>((set) => ({
+  isWelcomeComponentOpen: false,
+  handleClickOnNewDesign: () => {
+    set(() => ({
+      isWelcomeComponentOpen: true,
+    }));
+  },
+}));
